+++ packages/react/src/experimental/Navigation/Sidebar/Footer/index.tsx
@@ -24,7 +24,7 @@ export function SidebarFooter({
   user,
   options,
   showActivityButton = false,
-  activityButtonShortcut = ["cmd", "B"],
+  activityButtonShortcut,
   onActivityButtonClick,
   hasActivityUpdates,
 }: SidebarFooterProps) {