+++ packages/react/src/experimental/RichText/CoreEditor/Extensions/Color/index.tsx
@@ -1,0 +1,3 @@
+import Color from "@tiptap/extension-color"
+
+export const ColorExtension = Color